10/25/2020 Hv Changes - Damage Tab -- v2
missed 2 numbers to change.

Visually it is working now.

If you click a checkbox the other is disabled.  But the x stays in both boxes until you reload the menu.

You can uncheck and check boxes and it visually works now compared to previous versions.  

Im ok with keep this as is, unless someone else can add that final polish but its not needed IMO.