Hv Fixes - 2/6/2021
Menu added, ( Not functional ), Both sniper and LG show in f3 menu.  F3 Menu is modified with a new image to show all the options that have been added.

TODO:
Make a var that works and removes either Lightning / ClassicSniper from player inventory.  Once this is done we can release.